Feature / URL Shortener
Adds HashCacheService

+++ src/main/java/faang/school/urlshortenerservice/config/UrlShortenerProperties.java
@@ -6,7 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @Builder
 @ConfigurationProperties(prefix = "url-shortener")
 public record UrlShortenerProperties(String hostName,
-                                     long localCacheCapacity,
+                                     int localCacheCapacity,
                                      long hashAmountToLocalCache,
                                      double localCacheThresholdRatio,
                                      long hashAmountToGenerate,
